+++ src/Middleware/ImplicitOptionsMiddleware.php
@@ -1,0 +1,79 @@
+<?php
+/**
+ * @see       https://github.com/zendframework/zend-expressive-router for the canonical source repository
+ * @copyright Copyright (c) 2018 Zend Technologies USA Inc. (http://www.zend.com)
+ * @license   https://github.com/zendframework/zend-expressive-router/blob/master/LICENSE.md New BSD License
+ */
+
+namespace Zend\Expressive\Router\Middleware;
+
+use Fig\Http\Message\RequestMethodInterface as RequestMethod;
+use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\ServerRequestInterface;
+use Webimpress\HttpMiddlewareCompatibility\HandlerInterface as RequestHandlerInterface;
+use Webimpress\HttpMiddlewareCompatibility\MiddlewareInterface;
+use Zend\Expressive\Router\RouteResult;
+
+use const Webimpress\HttpMiddlewareCompatibility\HANDLER_METHOD;
+
+/**
+ * Handle implicit OPTIONS requests.
+ *
+ * Place this middleware after the routing middleware so that it can handle
+ * implicit OPTIONS requests: requests where OPTIONS is used, but the route
+ * does not explicitly handle that request method.
+ *
+ * When invoked, it will create a response with status code 200 and an Allow
+ * header that defines all accepted request methods.
+ *
+ * You may optionally pass a response prototype to the constructor; when
+ * present, that prototype will be used to create a new response with the
+ * Allow header.
+ *
+ * The middleware is only invoked in these specific conditions:
+ *
+ * - an OPTIONS request
+ * - with a `RouteResult` present
+ * - where the `RouteResult` contains a `Route` instance
+ * - and the `Route` instance defines implicit OPTIONS.
+ *
+ * In all other circumstances, it will return the result of the delegate.
+ */
+class ImplicitOptionsMiddleware implements MiddlewareInterface
+{
+    /**
+     * @var ResponseInterface
+     */
+    private $response;
+
+    public function __construct(ResponseInterface $response)
+    {
+        $this->response = $response;
+    }
+
+    /**
+     * Handle an implicit OPTIONS request.
+     *
+     * @return ResponseInterface
+     */
+    public function process(ServerRequestInterface $request, RequestHandlerInterface $handler)
+    {
+        if ($request->getMethod() !== RequestMethod::METHOD_OPTIONS) {
+            return $handler->{HANDLER_METHOD}($request);
+        }
+
+        $result = $request->getAttribute(RouteResult::class);
+        if (! $result) {
+            return $handler->{HANDLER_METHOD}($request);
+        }
+
+        $route = $result->getMatchedRoute();
+        if (! $route || ! $route->implicitOptions()) {
+            return $handler->{HANDLER_METHOD}($request);
+        }
+
+        $methods = implode(',', $route->getAllowedMethods());
+
+        return $this->response->withHeader('Allow', $methods);
+    }
+}